pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-macro-mcp"
version = "0.2.0"
description = "MCP server for macro-economic research — FRED, ECB, Eurostat, BIS, IMF, OECD"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27",
    "pandas>=2.1",
    "numpy>=1.26",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
]

[project.scripts]
claude-macro-mcp = "src.server:main"
